feat(mosaicod): Add new test for do_put with schema and empty data [#482]
- Added new test for do_put, loading empty record batch with schema
- Splitted tests in different files
